delete category button fixed
